Change "client" to "SDK" (#149)
## Problem

We are inconsistent in how we refer to our client libraries.

## Solution

As we mostly refer to our "SDKs" rather than our "clients", we've
updated our docs to prefer "SDK". This PR makes equivalent updates to
the Java SDK README.

Addresses: https://app.asana.com/0/1206419570114940/1207978475519546/f

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.