Add embed endpoint (#153)
## Problem

Add embed endpoint for Java SDK to allow users to create embeddings for
text data such as passage or query using a specified model. More details
on inference api can be found
[here](https://docs.pinecone.io/guides/inference/understanding-inference).

## Solution

Users can now call embed endpoint with the following parameters:
1. `String model`: Accepts a string from the specified
[models](https://docs.pinecone.io/models/overview) .
2. `Map<String, Object> parameters`: Accepts `input_type` and `truncate`
as keys with their corresponding values in a Map. The values are
expected to be scalar. Please note that the default value of `truncate`
is set to `END` if not specified.
3. `List<String> inputs`: The list must be of size atleast 1.

As a part of this change, I have added a `getInferenceClient()` in
`Pinecone` class to follow a similar pattern with other SDKs. The
underlying client uses `OkHTTPClient` for REST calls. The method
`getInferenceClient()` returns an instance of `Inference` class which I
added as a wrapper to the Inference API. So far, this wrapper contains
`embed()` endpoint only, with the plan of adding `rerank()` soon.
Lastly, I have added docstrings and updated README with an example of
the embed endpoint.

## Type of Change

- [X] New feature (non-breaking change which adds functionality)

## Test Plan

Added integration tests.