Generate code based on 2024-10 api spec (#158)
## Problem

Generate code based on 2024-10 api spec.

## Solution
OpenAPI based code was regenerated using the 2024-10 api sec which
resulted in the following changes:
1. Updated `build-oas.sh` to include all control plane, data plane, and
inference modules.
2. Inference is now broken out of control plane so the underlying
`inference` code generated by openAPI spec is now updated. This also
resulted in instantiating a separate `ApiClient` which is a part of
inference module now instead of the shared `ApiClient` class of control
plane. Also as a part of this refactoring, the `customOkHttpClient` is
now a part of the `PineconeConfig` class, so once the
`customOkHttpClient` is set by the user at the time of instantiating
`Pinecone` using its builder class, it'll be stored in the
`PineconeConfig` instance which can further be leveraged by the
`Inference` api calls as well.
3. Control and data plane modules are now prepended with `db_` so this
resulted in updating import statements for a lot of existing classes.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Since no new functionality is added, the existing integration tests
should run fine.