Update README based on regenerated code (#164)
## Problem

Regenerating the code based on new OAS caused changes in imports, so the
examples in the README needs to be updated accordingly.

## Solution

Changes include: 
1. Updating import statements since inference is now broken out of
control, so the paths for the openAPI generated model classes like
`ApiException`, `IndexModel`, etc. were updated.
2. Import statement was missing for an example or two

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

No code changes so integration tests should run successfully.