Create new config per connection (#178)
## Problem

When connecting to multiple indexes using the same Pinecone instance,
the connections for each index are cached in a ConcurrentHashMap called
connectionsMap, which maps an indexName to its corresponding
PineconeConnection. However, after establishing connections for multiple
indexes, all entries in the map incorrectly point to the most recently
created PineconeConnection instead of maintaining separate connections
for each index.

## Solution

When assigning a host to the PineconeConfig object, which is passed as
input for establishing a gRPC connection, create a new PineconeConfig
instance instead of modifying the existing one. Otherwise, all
connection objects stored in the connectionsMap will share the same
PineconeConfig instance, causing them to reference the most recently
created PineconeConnection instead of maintaining separate
configurations for each index.

## Type of Change

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Added integration test.