Generate core based on oas and protos (#204)
## Problem

Update core based on 2025-10 specs

## Solution

Generated and updated the core based on the 2025-10 OAS and protos. Note
that the code had already been generated from the 2025-10 specs, but
some fixes were made to the OAS, so I had to regenerate it.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

No new functionalities are added in this pr so existing integration
tests should pass.